style: cast int/float operand pairs explicitly (Psalm 7 strict binary operands)
Where an arithmetic operation mixed an int with a float, both operands are now cast to
float (an int literal becomes a float literal). The result already was a float, so no value
changes. InvalidOperand 'ints and floats' 816 -> 288, total 23,092 -> 22,543.

+++ gl/accruals.php
@@ -64,7 +64,7 @@ if (isset($_POST['go']) || isset($_POST['show']))
 	if ($input_error == 0)
 	{
 		$periods = input_num('periods');
-		$per = $periods - 1;
+		$per = (float)$periods - 1.0;
 		$date = $date_ = get_post('date_');
 		$freq = get_post('freq');
 		if ($freq == 3 || $freq == 4) {
@@ -72,10 +72,10 @@ if (isset($_POST['go']) || isset($_POST['show']))
 			$date  = end_month($date_); // avoid skip on shorter months
 		}
 		
-		$lastdate = ($freq == 1 ? add_days($date_, 7*$per) :
-			($freq == 2 ? add_days($date_, 14*$per) :
+		$lastdate = ($freq == 1 ? add_days($date_, 7.0*(float)$per) :
+			($freq == 2 ? add_days($date_, 14.0*(float)$per) :
 			($freq == 3 ? end_month(add_months($date_, $per)) : 
-			end_month(add_months($date_, 3*$per)))));
+			end_month(add_months($date_, 3.0*(float)$per)))));
 		if (!is_date_in_fiscalyears($lastdate, false))
 		{
 			display_error(_("Some of the period dates are outside the fiscal year or are closed for further data entry. Create a new fiscal year first!"));
@@ -85,9 +85,9 @@ if (isset($_POST['go']) || isset($_POST['show']))
 		if ($input_error == 0)
 		{
 			$amount = input_num('amount');
-			$am = round2($amount / $periods, user_price_dec());
-			if ($am * $periods != $amount)
-				$am0 = $am + $amount - $am * $periods;
+			$am = round2((float)$amount / (float)$periods, user_price_dec());
+			if ((float)$am * (float)$periods != $amount)
+				$am0 = (float)$am + (float)$amount - (float)$am * (float)$periods;
 			else
 				$am0 = $am;
 			if (get_post('memo_') != "")
@@ -159,7 +159,7 @@ if (isset($_POST['go']) || isset($_POST['show']))
 						label_cell("");
 					if ($dim > 1)
 						label_cell("");
-					display_debit_or_credit_cells($am0 * -1);
+					display_debit_or_credit_cells((float)$am0 * (float)(-1));
 					label_cell($memo);
 					alt_table_row_color($k);
 					label_cell($date);
