+++ sales/sales_order_entry.php
@@ -577,7 +577,7 @@ function handle_update_item(): void
 	if ($_POST['UpdateItem'] != '' && check_item_data()) {
 		session_obj('Items')->update_cart_item($_POST['LineNo'],
 		 input_num('qty'), input_num('price'),
-		 input_num('Disc') / 100, $_POST['item_description'] );
+		 (float)input_num('Disc') / 100.0, $_POST['item_description'] );
 	}
 	page_modified();
   line_start_focus();
@@ -604,7 +604,7 @@ function handle_new_item(): void
 			return;
 	}
 	add_to_order($_SESSION['Items'], get_post('stock_id'), input_num('qty'),
-		input_num('price'), input_num('Disc') / 100, get_post('stock_id_text'));
+		input_num('price'), (float)input_num('Disc') / 100.0, get_post('stock_id_text'));
 
 	unset($_POST['_stock_id_edit'], $_POST['stock_id']);
 	page_modified();
